refactor: moved other widgets into separate files
