+++ lib/dialogs/sell/sell_dialog.dart
@@ -468,105 +468,3 @@ class ShortAptButton extends StatelessWidget {
     );
   }
 }
-
-class AptIcon extends StatelessWidget {
-  const AptIcon({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    final aptTypeSelection =
-        context.select((SellDialogBloc bloc) => bloc.state.aptTypeSelection);
-    return Container(
-      width: 35,
-      height: 35,
-      decoration: BoxDecoration(
-        shape: BoxShape.circle,
-        image: DecorationImage(
-          scale: 0.5,
-          image: aptTypeSelection.isLong
-              ? const AssetImage(
-                  'assets/images/apt_noninverted.png',
-                )
-              : const AssetImage(
-                  'assets/images/apt_inverted.png',
-                ),
-        ),
-      ),
-    );
-  }
-}
-
-class Price extends StatelessWidget {
-  const Price({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: [
-        Text(
-          'Price:',
-          style: textStyle(
-            Colors.white,
-            15,
-            isBold: false,
-            isUline: false,
-          ),
-        ),
-        BlocBuilder<SellDialogBloc, SellDialogState>(
-          buildWhen: (previous, current) =>
-              previous.aptTypeSelection != current.aptTypeSelection ||
-              previous.aptSellInfo != current.aptSellInfo ||
-              previous.longApt != current.longApt ||
-              previous.shortApt != current.shortApt,
-          builder: (context, state) {
-            final price = state.aptSellInfo.axPrice.toStringAsFixed(6);
-            final _textStyle = textStyle(
-              Colors.white,
-              15,
-              isBold: false,
-              isUline: false,
-            );
-            return state.aptTypeSelection.isLong
-                ? Text(
-                    '$price AX per ${state.longApt.ticker} APT',
-                    style: _textStyle,
-                  )
-                : Text(
-                    '$price AX per ${state.shortApt.ticker} APT',
-                    style: _textStyle,
-                  );
-          },
-        )
-      ],
-    );
-  }
-}
-
-class Ticker extends StatelessWidget {
-  const Ticker({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocBuilder<SellDialogBloc, SellDialogState>(
-      buildWhen: (previous, current) =>
-          previous.aptTypeSelection != current.aptTypeSelection ||
-          previous.longApt != current.longApt ||
-          previous.shortApt != current.shortApt,
-      builder: (context, state) {
-        final _textStyle = textStyle(
-          Colors.white,
-          15,
-          isBold: false,
-          isUline: false,
-        );
-        return Text(
-          state.aptTypeSelection.isLong
-              ? '''${state.longApt.ticker} APT'''
-              : '''${state.shortApt.ticker} APT''',
-          style: _textStyle,
-        );
-      },
-    );
-  }
-}
