+++ lib/dialogs/sell/widgets/widgets.dart
@@ -1,6 +1,9 @@
+export 'apt_icon.dart';
 export 'balance.dart';
 export 'market_price_impact.dart';
 export 'minimum_received.dart';
+export 'price.dart';
 export 'slippage_tolerance.dart';
+export 'ticker.dart';
 export 'total_fees.dart';
 export 'you_received.dart';
